fix: prevent workflow stutter due to point confusion on forge visits
Our filter/render code tries to magically reposition to the last notification
at point prior to a re-render. When triggering buffer switches you lose the
active point and this can result in point confusion where the point lags in
the wrong place in the gh-notify buffer. We work around this by simply
pre-rendering for the notification visit into forge.