hack: work around magit git repo restrictions for forge-visit
magit wants to live inside a git repo context to be able to spin up issue and
pull request buffers for topics from the forge ... we don't actually need to
have local checkouts of a repo to interact with the issues/pr's for
the most part ... i.e. you can even do a full web based review through
github-review from a template buffer with no real local backing repo ... but
we run into a nag if it can't find the repository locally and as a final
attempt magit tries to operate from default-directory

so until we find something cleaner, we just make a tmp empty git repo and let
that be the default directory for forge-visit ... that saves us from having to
go and clone a bunch of crap we don't need locally on our systems just to
reply to comments and do light-reviews.