doc, feature: added issue/pullreq listing for repo at point
I noticed I often want a way to directly go to another issue or PR
associated to the repo for the notification at point without having to visit
the magit status for the repo first.

Added completing-read based navigation to issues and pull requests for the
repo associated to the notification at point. Bound to C-i and C-p respectively.

reformatted mode docs a bit for cosmetics.