v2.0.0
This version of gh-notify is only compatible with the latest version
of Magit/Forge as of April, 2024.

It leverages Forge's new built-in incremental fetch logic as well as
its updated unread/read state management.

I've updated this for folks who are running the latest versions of
Magit/Forge yet still prefer the gh-notify UX for filtering/sorting
notifications.

If you are running older versions of Magit/Forge pre October 2023
please use v0.1.0 of gh-notify.